TOOLS/lua/autoload: add ignore_hidden option
In 3c7f95d3061e29c2dc43da57f1ad2b6d356e5b83 files that start with a '.'
were blacklisted from autoload.lua. Since then
5c1ed54a3372a2d5bdbec4aecdfab56d908e7718 was introduced and explicitly
whitelisted file extensions. With this change, there is no longer a
reason to blacklist all files starting with '.', because it is valid to
have a file called '.hidden.mkv', and there is no chance of hidden files
such as '.bashrc', '.DS_STORE', '.gitignore', etc. being autoloaded by
mpv.

This commit tries to keep the same behavior as before, which is to by
default not load hidden files, but allows the user to optionally allow
hidden files to be autoloaded.